created return all items and return total functionality

+++ TacoLoco/src/main/java/com/example/TacoLoco/model/Item.java
@@ -11,11 +11,11 @@ public class Item {
 
     private final String name;
 
-    private final BigDecimal price;
+    private final double price;
 
     private final int quantity;
 
-    public Item(@JsonProperty("id") UUID id, @JsonProperty("name") String name, @JsonProperty("price") BigDecimal price, @JsonProperty("quantity") int quantity){
+    public Item(@JsonProperty("id") UUID id, @JsonProperty("name") String name, @JsonProperty("price") double price, @JsonProperty("quantity") int quantity){
         this.id = id;
         this.name = name;
         this.price = price;
@@ -30,7 +30,7 @@ public class Item {
         return name;
     }
     
-    public BigDecimal getPrice(){
+    public double getPrice(){
         return price;
     }
